requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose
jinja2